Force upload trace with 1st capture position
* Bugfix: handle expired token

+++ App.js
@@ -1,19 +1,14 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 import Icon from 'react-native-vector-icons/Ionicons';
+import { Platform } from 'react-native';
 import AsyncStorage from '@react-native-community/async-storage';
-import {NavigationContainer} from '@react-navigation/native';
-import {createStackNavigator} from '@react-navigation/stack';
-// import {ApolloClient, HttpLink, ApolloLink, concat} from 'apollo-client';
-import {ApolloProvider} from '@apollo/react-hooks';
-import HomeTabs from './screens/HomeTabs';
-import SplashScreen from './screens/Splash';
-import SignInScreen from './screens/SignIn';
+import { ApolloProvider } from '@apollo/react-hooks';
 import apolloClient from './libs/apollo';
+import Navi from './Navigation'
 
 Icon.loadFont();
 
-const Stack = createStackNavigator();
 export const AuthContext = React.createContext();
 
 const App = () => {
@@ -62,7 +57,7 @@ const App = () => {
 
       // This will switch to the App screen or Auth screen and this loading
       // screen will be unmounted and thrown away.
-      dispatch({type: 'RESTORE_TOKEN', token: userToken});
+      dispatch({ type: 'RESTORE_TOKEN', token: userToken });
     };
 
     bootstrapAsync();
@@ -77,24 +72,34 @@ const App = () => {
         // In the example, we'll use a dummy token
         const keys = Object.keys(data);
         if (keys.includes('errors')) {
-          AsyncStorage.clear();
+          if (Platform.OS === 'android') {
+            await AsyncStorage.clear();
+          }
+          if (Platform.OS === 'ios') {
+            await AsyncStorage.getAllKeys().then(AsyncStorage.multiRemove)
+          }
         } else {
           AsyncStorage.setItem('userToken', data.token);
           AsyncStorage.setItem('userResp', JSON.stringify(data));
-          dispatch({type: 'SIGN_IN', token: data.token});
+          dispatch({ type: 'SIGN_IN', token: data.token });
         }
+        // console.log('data == > ', data)
       },
-      signOut: () => {
-        AsyncStorage.clear();
-        dispatch({type: 'SIGN_OUT'});
+      signOut: async () => {
+        if (Platform.OS === 'android') {
+          await AsyncStorage.clear();
+        }
+        if (Platform.OS === 'ios') {
+          await AsyncStorage.getAllKeys().then(AsyncStorage.multiRemove)
+        }
+        dispatch({ type: 'SIGN_OUT' });
       },
       signUp: async data => {
         // In a production app, we need to send user data to server and get a token
         // We will also need to handle errors if sign up failed
         // After getting token, we need to persist the token using `AsyncStorage`
         // In the example, we'll use a dummy token
-
-        dispatch({type: 'SIGN_IN', token: ''});
+        dispatch({ type: 'SIGN_IN', token: '' });
       },
     }),
     [],
@@ -103,30 +108,11 @@ const App = () => {
   return (
     <AuthContext.Provider value={authContext}>
       <ApolloProvider client={apolloClient}>
-        <NavigationContainer>
-          <Stack.Navigator
-            screenOptions={{
-              headerShown: false,
-            }}>
-            {state.isLoading ? (
-              // We haven't finished checking for the token yet
-              <Stack.Screen name="Splash" component={SplashScreen} />
-            ) : state.userToken == null ? (
-              // No token found, user isn't signed in
-              <Stack.Screen
-                name="SignIn"
-                component={SignInScreen}
-                options={{
-                  title: 'Sign in',
-                  // When logging out, a pop animation feels intuitive
-                  animationTypeForReplace: state.isSignout ? 'pop' : 'push',
-                }}
-              />
-            ) : (
-              <Stack.Screen name="Home" component={HomeTabs} />
-            )}
-          </Stack.Navigator>
-        </NavigationContainer>
+        <Navi
+          isLoading={state.isLoading}
+          userToken={state.userToken}
+          isSignout={state.isSignout}
+        />
       </ApolloProvider>
     </AuthContext.Provider>
   );
